fastapi/nlp-service/requirements.txt
fastapi
uvicorn
pymysql
sqlalchemy
aiomysql
python-dotenv
openai
httpx
cryptography
pytz
pandas
yt-dlp
torch
transformers
pyannote.audio